Update XNNPACK backend utilities to use `input_spec` node metadata for
identifying parameters and retrieving tensor data, with fallback to
graph signature for backward compatibility.

Differential Revision: [D95876982](https://our.internmc.facebook.com/intern/diff/D95876982/)

[ghstack-poisoned]

+++ backends/qualcomm/aot/wrappers/TensorWrapper.h
@@ -76,7 +76,7 @@ class TensorWrapper {
           rank);
       return;
     }
-    for (int i = 0; i < rank; ++i) {
+    for (size_t i = 0; i < rank; ++i) {
       QNN_TENSOR_VER_PTR(tensor_)->dimensions[i] = dims[i];
     }
   }
